+++ benchit/__init__.py
@@ -1,6 +1,6 @@
-from benchit.bench import BenchmarkObj, timings, _get_possible_indexbys
+from benchit.main import BenchmarkObj, timings, bench, _get_possible_indexbys
 from benchit.plot_utils import extract_modules_from_globals, print_specs
 
-__version__ = '0.0.1'
+__version__ = '0.0.2'
 
-__all__ = ["BenchmarkObj", "timings", "extract_modules_from_globals", "print_specs"]
+__all__ = ["BenchmarkObj", "bench", "timings", "extract_modules_from_globals", "print_specs"]