+++ docs/source/conf.py
@@ -25,7 +25,7 @@ author = u'Divakar Roy'
 # The short X.Y version
 version = u''
 # The full version, including alpha/beta/rc tags
-release = u'0.0.1'
+release = u'0.0.2'
 
 
 # -- General configuration ---------------------------------------------------
@@ -37,7 +37,7 @@ release = u'0.0.1'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.napoleon', 'sphinx.ext.autodoc']
+extensions = ['sphinx.ext.napoleon', 'sphinx.ext.autodoc', 'sphinx.ext.autosectionlabel']
 
 
 # Add any paths that contain templates here, relative to this directory.